more printf() format warnings fixed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_4_BRANCH@17231 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/unix/mimetype.cpp
@@ -2367,11 +2367,11 @@ bool wxMimeTypesManagerImpl::ReadMailcap(const wxString& strFileName,
                             // something is wrong, bail out
                             cont = FALSE;
 
-                            wxLogDebug(wxT("Mailcap file %s, line %d: "
+                            wxLogDebug(wxT("Mailcap file %s, line %lu: "
                                            "'\\' on the end of the last line "
                                            "ignored."),
                                        strFileName.c_str(),
-                                       nLine + 1);
+                                       (unsigned long)nLine + 1);
                         }
                         else
                         {
@@ -2433,11 +2433,11 @@ bool wxMimeTypesManagerImpl::ReadMailcap(const wxString& strFileName,
                                 // this might be useful for the programmer
                                 wxLogDebug
                                 (
-                                    wxT("Mailcap file %s, line %d: "
+                                    wxT("Mailcap file %s, line %lu: "
                                         "unknown field '%s' for the "
                                         "MIME type '%s' ignored."),
                                     strFileName.c_str(),
-                                    nLine + 1,
+                                    (unsigned long)nLine + 1,
                                     curField.c_str(),
                                     data.type.c_str()
                                 );
